:truck: rename: global/security 로 파일 위치 이동 #16
